refactor: hook 폴더에 있던 파일과 utils에 있는 파일을 관심사에 맞게 분리하고 이름 변경
util에 있던 내장 메서드를 사용하는 함수들 불필요하다고 판단하여 제거
