+++ docs/REFACTOR_2.md
@@ -9,6 +9,6 @@
 - [x] Button 컴포넌트, Modal 컴포넌트 등 범용적으로 사용하는 컴포넌트를 레퍼런스(Antd, Chark) 참고하여 더 고도화 시키기
 - [x] 모든 z-index 상수로 관리하기
 - [x] 오브젝트를 상수로 선언할 때 키값을 대문자로 변경하기 (상수라는 것을 더 명확히 표시하기 위해서)
-- [ ] hook 폴더에 있는 파일들, utils에 있는 파일들을 관심사에 맞게 분리하고 이름을 함수의 동작과 연관지어 짓기
-- [ ] 내장 메서드를 사용하는 util 함수가 필요한 지 고민해보기
+- [x] hook 폴더에 있는 파일들, utils에 있는 파일들을 관심사에 맞게 분리하고 이름을 함수의 동작과 연관지어 짓기
+- [x] 내장 메서드를 사용하는 util 함수가 필요한 지 고민해보기
 - [x] 영어 오탈자 고치기
