+++ src/components/common/Input/Input.tsx
@@ -1,6 +1,6 @@
 import { ChangeEvent, InputHTMLAttributes, KeyboardEvent } from 'react';
 import styled from 'styled-components';
-import { setNextInputFocus } from '../../../utils';
+import { setNextInputFocus } from '../../../utils/common';
 
 export interface InputProps extends InputHTMLAttributes<HTMLInputElement> {
   textAlign?: string;
